refactor(rsc): extract PPR stream utilities
