+++ packages/plugin-rsc/examples/ppr/src/framework/stream-utils.ts
@@ -1,0 +1,27 @@
+export function stringToStream(data: string): ReadableStream<Uint8Array> {
+  return new ReadableStream({
+    start(controller) {
+      controller.enqueue(new TextEncoder().encode(data))
+      controller.close()
+    },
+  })
+}
+
+export function concatStreams(
+  first: ReadableStream<Uint8Array>,
+  second: ReadableStream<Uint8Array>,
+): ReadableStream<Uint8Array> {
+  return first.pipeThrough(
+    new TransformStream<Uint8Array, Uint8Array>({
+      async flush(controller) {
+        await second.pipeTo(
+          new WritableStream({
+            write(chunk) {
+              controller.enqueue(chunk)
+            },
+          }),
+        )
+      },
+    }),
+  )
+}
